gcloud/requirements.txt
google-cloud-speech==2.13.1
google-cloud-storage==2.1.0
tqdm==4.63.0